Organized directory structure
(somewhat) cleaned up the directory structure. Grouped related files:
- Added `src/components/Diagram` for Diagram-related components
- Added `src/components/Displays` for files related to Diagram Element Displays

Still need to sort through all `src/lib/` files to see what can be consolidated and organized there,
but this is a start.
