+++ src/components/Diagram/Displays/DisplayTypeRegistry.tsx
