+++ src/components/Diagram/Displays/DisplaysSection.tsx
