+++ src/components/Diagram/Displays/displayTypes/ControlBoolean.tsx
