+++ src/components/Diagram/Displays/displayTypes/ControlRange.tsx
