+++ src/components/Diagram/Displays/displayTypes/ControlText.tsx
